Add entrypoint script to Dockerfile to start Django application
This commit adds an entrypoint script to the Dockerfile that starts the Django application. The entrypoint script is copied into the image and set as the entry point. The script starts the Django application using the 'python manage.py runserver 0.0.0.0:8000' command. This change enables the application to be started automatically when the Docker container is run.